monthly cleanup

+++ ft_printf/len_functions.c
@@ -1,8 +1,8 @@
 #include "ft_printf.h"
 
-size_t	unsigned_len(unsigned long long num)
+unsigned int	unsigned_len(unsigned long long num)
 {
-	size_t	num_len;
+	unsigned int	num_len;
 
 	num_len = 1;
 	while (num / ft_unsigned_pow(10, num_len - 1) >= 10)
@@ -16,9 +16,10 @@ static int	ft_abs(int nb)
 		return (nb * -1);
 	return (nb);
 }
-size_t	num_len(int	num)
+
+unsigned int	num_len(int	num)
 {
-	size_t	n_len;
+	unsigned int	n_len;
 
 	n_len = 1;
 	if (num == -2147483648)
